requirements.txt
langgraph>=0.4.8
langchain[openai]>=0.3.25
python-dotenv>=1.0.0
langchain-mcp-adapters>=0.1.0
mcp>=1.0.0
aiohttp>=3.9.0
nest-asyncio>=1.5.0
requests>=2.31.0 